requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.8
pytz==2019.3
requests==2.22.0
selenium==3.141.0
soupsieve==1.9.5
sqlparse==0.3.0
tmdbv3api==1.5.1
urllib3==1.25.7
